# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Swagger Docs",
    "description": "Docs for app backend",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "host": "localhost:3003",
  "basePath": "/api",
...
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/item": {
      "post": {
        "tags": [
          "Items"
        ],
        "description": "Create new item in system",
        "parameters": [
          {
            "name": "item",
            "in": "body",
            "description": "Item that we want to create",
            "schema": {
              "$ref": "#/definitions/Item"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "New item is created",
            "schema": {
              "$ref": "#/definitions/Item"
            }
          }
        }
      },
      "get": {
        "tags": [
          "Items"
        ],
        "summary": "Get all items in system",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Items"
            }
          }
        }
      }
    },
    "/item/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "description": "ID of item that we want to find",
          "type": "string"
        }
      ],
      "get": {
        "tags": [
          "Items"
        ],
        "summary": "Get item with given ID",
        "responses": {
          "200": {
            "description": "Item is found",
            "schema": {
              "$ref": "#/definitions/Item"
            }
          }
        }
      },
      "delete": {
        "summary": "Delete item with given ID",
        "tags": [
          "Items"
        ],
        "responses": {
          "200": {
            "description": "Item is deleted"
          }
        }
      }
    }
  },
  "definitions": {
...
    "Item": {
      "required": [
        "name"
      ],
      "properties": {
        "name": {
          "type": "string"
        },
        "cuisine": {
          "type": "string"
        },
        "isVegetarian": {
          "type": "boolean"
        },
        "image": {
          "type": "string"
        },
        "price": {
          "type": "string"
        },
        "label": {
          "type": "string"
        },
        "rating": {
          "type": "number"
        }
      }
    },
    "Items": {
      "type": "array",
      "$ref": "#/definitions/Item"
    },
...
  }
}